requirements.txt
fastapi
uvicorn[standard]
pydantic
mysql-connector-python
strawberry-graphql[debug-server]
strawberry-graphql[sqlalchemy]
sqlalchemy==2.0.6